Successfully clicks through most queues
Still a small problem with sequential age verification pages, but it should be an easy fix when I give it a fresh set of eyes. Need to test changing from one queue to another, and also might add a fancy ending message to the console to ensure the user knows that the program has finished. Should probably also end Nightmare at that point and kill the program.